Fix CUDA/OpenMP compiler warnings/errors on Windows

+++ src/template/compress.c
@@ -10,12 +10,12 @@ _t2(compress, Scalar, 1)(zfp_stream* stream, const zfp_field* field)
 
   /* compress array one block of 4 values at a time */
   for (x = 0; x < mx; x += 4, data += 4) {
-    uint16 block_size = _t2(zfp_encode_block, Scalar, 1)(stream, data);
+    uint16 block_size = (uint16)_t2(zfp_encode_block, Scalar, 1)(stream, data);
     if (length_table)
       *length_table++ = block_size;
   }
   if (x < nx) {
-    uint16 block_size = _t2(zfp_encode_partial_block_strided, Scalar, 1)(stream, data, nx - x, 1);
+    uint16 block_size = (uint16)_t2(zfp_encode_partial_block_strided, Scalar, 1)(stream, data, nx - x, 1);
     if (length_table)
       *length_table++ = block_size;
   }
@@ -36,9 +36,9 @@ _t2(compress_strided, Scalar, 1)(zfp_stream* stream, const zfp_field* field)
     const Scalar* p = data + sx * (ptrdiff_t)x;
     uint16 block_size;
     if (nx - x < 4)
-      block_size = _t2(zfp_encode_partial_block_strided, Scalar, 1)(stream, p, nx - x, sx);
+      block_size = (uint16)_t2(zfp_encode_partial_block_strided, Scalar, 1)(stream, p, nx - x, sx);
     else
-      block_size = _t2(zfp_encode_block_strided, Scalar, 1)(stream, p, sx);
+      block_size = (uint16)_t2(zfp_encode_block_strided, Scalar, 1)(stream, p, sx);
     if (length_table)
       *length_table++ = block_size;
   }
@@ -62,9 +62,9 @@ _t2(compress_strided, Scalar, 2)(zfp_stream* stream, const zfp_field* field)
       const Scalar* p = data + sx * (ptrdiff_t)x + sy * (ptrdiff_t)y;
       uint16 block_size;
       if (nx - x < 4 || ny - y < 4)
-        block_size = _t2(zfp_encode_partial_block_strided, Scalar, 2)(stream, p, MIN(nx - x, 4u), MIN(ny - y, 4u), sx, sy);
+        block_size = (uint16)_t2(zfp_encode_partial_block_strided, Scalar, 2)(stream, p, MIN(nx - x, 4u), MIN(ny - y, 4u), sx, sy);
       else
-        block_size = _t2(zfp_encode_block_strided, Scalar, 2)(stream, p, sx, sy);
+        block_size = (uint16)_t2(zfp_encode_block_strided, Scalar, 2)(stream, p, sx, sy);
       if (length_table)
         *length_table++ = block_size;
     }
@@ -91,9 +91,9 @@ _t2(compress_strided, Scalar, 3)(zfp_stream* stream, const zfp_field* field)
         const Scalar* p = data + sx * (ptrdiff_t)x + sy * (ptrdiff_t)y + sz * (ptrdiff_t)z;
         uint16 block_size;
         if (nx - x < 4 || ny - y < 4 || nz - z < 4)
-          block_size = _t2(zfp_encode_partial_block_strided, Scalar, 3)(stream, p, MIN(nx - x, 4u), MIN(ny - y, 4u), MIN(nz - z, 4u), sx, sy, sz);
+          block_size = (uint16)_t2(zfp_encode_partial_block_strided, Scalar, 3)(stream, p, MIN(nx - x, 4u), MIN(ny - y, 4u), MIN(nz - z, 4u), sx, sy, sz);
         else
-          block_size = _t2(zfp_encode_block_strided, Scalar, 3)(stream, p, sx, sy, sz);
+          block_size = (uint16)_t2(zfp_encode_block_strided, Scalar, 3)(stream, p, sx, sy, sz);
         if (length_table)
           *length_table++ = block_size;
       }
@@ -123,9 +123,9 @@ _t2(compress_strided, Scalar, 4)(zfp_stream* stream, const zfp_field* field)
           const Scalar* p = data + sx * (ptrdiff_t)x + sy * (ptrdiff_t)y + sz * (ptrdiff_t)z + sw * (ptrdiff_t)w;
           uint16 block_size;
           if (nx - x < 4 || ny - y < 4 || nz - z < 4 || nw - w < 4)
-            block_size = _t2(zfp_encode_partial_block_strided, Scalar, 4)(stream, p, MIN(nx - x, 4u), MIN(ny - y, 4u), MIN(nz - z, 4u), MIN(nw - w, 4u), sx, sy, sz, sw);
+            block_size = (uint16)_t2(zfp_encode_partial_block_strided, Scalar, 4)(stream, p, MIN(nx - x, 4u), MIN(ny - y, 4u), MIN(nz - z, 4u), MIN(nw - w, 4u), sx, sy, sz, sw);
           else
-            block_size = _t2(zfp_encode_block_strided, Scalar, 4)(stream, p, sx, sy, sz, sw);
+            block_size = (uint16)_t2(zfp_encode_block_strided, Scalar, 4)(stream, p, sx, sy, sz, sw);
           if (length_table)
             *length_table++ = block_size;
         }
